fixed some visual artifacts

+++ src/lighting/Ray.java
@@ -11,7 +11,7 @@ public class Ray implements Cloneable {
     private Line2D line;
 
     public Ray(int x1, int y1, int x2, int y2){
-        line = new Line2D.Float(x1, y1, x2, y2);
+        line = new Line2D.Double(x1, y1, x2, y2);
     }
     public Ray(Line2D line) {
         this.line = line;
@@ -21,11 +21,11 @@ public class Ray implements Cloneable {
     public Ray increaseLength(double increase){
         double ndx = (increase * (line.getX2() - line.getX1())) / Utils.getLineLength(line),
                 ndy = (increase * (line.getY2() - line.getY1())) / Utils.getLineLength(line);
-        line = new Line2D.Float((int) line.getX1(), (int) line.getY1(), (int) (line.getX1() + ndx), (int) (line.getY1() + ndy));
+        line = new Line2D.Double((int) line.getX1(), (int) line.getY1(), (int) (line.getX1() + ndx), (int) (line.getY1() + ndy));
         return this;
     }
     public Ray rotate(double angle){
-        return new Ray(new Line2D.Float(
+        return new Ray(new Line2D.Double(
                 (int) line.getX1(),
                 (int) line.getY1(),
                 (int) (line.getX1() + Utils.getLineLength(line) * Math.cos(Math.toRadians(angle + getAngle()))),
